extract actions (setup/selectvalue)
https://coveord.atlassian.net/browse/KIT-633
